third_party/WebKit/LayoutTests/css-parser/color3-expected.txt

Color test for resources/css-parsing-tests/color3.json

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS colorTest.parseColor("") is "parse error"
PASS colorTest.parseColor(" /* hey */\n") is "parse error"
FAIL colorTest.parseColor("4") should be parse error. Was rgb(0, 0, 4).
PASS colorTest.parseColor("top") is "parse error"
PASS colorTest.parseColor("/**/transparent") is "rgba(0, 0, 0, 0)"
PASS colorTest.parseColor("transparent") is "rgba(0, 0, 0, 0)"
PASS colorTest.parseColor(" transparent\n") is "rgba(0, 0, 0, 0)"
PASS colorTest.parseColor("TransParent") is "rgba(0, 0, 0, 0)"
PASS colorTest.parseColor("currentColor") is "currentColor"
PASS colorTest.parseColor("CURRENTcolor") is "currentColor"
PASS colorTest.parseColor("current-Color") is "parse error"
PASS colorTest.parseColor("black") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("white") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("fuchsia") is "rgb(255, 0, 255)"
PASS colorTest.parseColor("cyan") is "rgb(0, 255, 255)"
PASS colorTest.parseColor("CyAn") is "rgb(0, 255, 255)"
PASS colorTest.parseColor("#") is "parse error"
PASS colorTest.parseColor("#f") is "parse error"
PASS colorTest.parseColor("#ff") is "parse error"
PASS colorTest.parseColor("#fff") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("#ffg") is "parse error"
PASS colorTest.parseColor("#ffff") is "parse error"
PASS colorTest.parseColor("#fffff") is "parse error"
PASS colorTest.parseColor("#ffffff") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("#fffffg") is "parse error"
PASS colorTest.parseColor("#fffffff") is "parse error"
PASS colorTest.parseColor("#ffffffff") is "parse error"
PASS colorTest.parseColor("#fffffffff") is "parse error"
PASS colorTest.parseColor("#FFCc99") is "rgb(255, 204, 153)"
PASS colorTest.parseColor("#369") is "rgb(51, 102, 153)"
PASS colorTest.parseColor("rgb(00, 51, 102)") is "rgb(0, 51, 102)"
PASS colorTest.parseColor("r\\gb(00, 51, 102)") is "rgb(0, 51, 102)"
PASS colorTest.parseColor("r\\67 b(00, 51, 102)") is "rgb(0, 51, 102)"
PASS colorTest.parseColor("RGB(153, 204, 255)") is "rgb(153, 204, 255)"
PASS colorTest.parseColor("rgB(0, 0, 0)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("rgB(0, 51, 255)") is "rgb(0, 51, 255)"
PASS colorTest.parseColor("rgb(0,51,255)") is "rgb(0, 51, 255)"
PASS colorTest.parseColor("rgb(0	,  51 ,255)") is "rgb(0, 51, 255)"
PASS colorTest.parseColor("rgb(/* R */0, /* G */51, /* B */255)") is "rgb(0, 51, 255)"
FAIL colorTest.parseColor("rgb(-51, 306, 0)") should be rgb(-51, 306, 0). Was rgb(0, 255, 0).
FAIL colorTest.parseColor("rgb(42%, 3%, 50%)") should be rgb(107, 7, 127). Was rgb(107, 7, 128).
PASS colorTest.parseColor("RGB(100%, 100%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("rgB(0%, 0%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("rgB(10%, 20%, 30%)") is "rgb(25, 51, 76)"
PASS colorTest.parseColor("rgb(10%,20%,30%)") is "rgb(25, 51, 76)"
PASS colorTest.parseColor("rgb(10%	,  20% ,30%)") is "rgb(25, 51, 76)"
PASS colorTest.parseColor("rgb(/* R */ 10%, /* G */ 20%, /* B */ 30%)") is "rgb(25, 51, 76)"
FAIL colorTest.parseColor("rgb(-12%, 110%, 1400%)") should be rgb(-31, 280, 3570). Was rgb(0, 255, 255).
PASS colorTest.parseColor("rgb(10%, 50%, 0)") is "parse error"
PASS colorTest.parseColor("rgb(255, 50%, 0%)") is "parse error"
PASS colorTest.parseColor("rgb(0, 0 0)") is "parse error"
PASS colorTest.parseColor("rgb(0, 0, 0deg)") is "parse error"
PASS colorTest.parseColor("rgb(0, 0, light)") is "parse error"
PASS colorTest.parseColor("rgb()") is "parse error"
PASS colorTest.parseColor("rgb(0)") is "parse error"
PASS colorTest.parseColor("rgb(0, 0)") is "parse error"
PASS colorTest.parseColor("rgb(0, 0, 0, 0)") is "parse error"
PASS colorTest.parseColor("rgb(0%)") is "parse error"
PASS colorTest.parseColor("rgb(0%, 0%)") is "parse error"
PASS colorTest.parseColor("rgb(0%, 0%, 0%, 0%)") is "parse error"
PASS colorTest.parseColor("rgb(0%, 0%, 0%, 0)") is "parse error"
PASS colorTest.parseColor("rgba(0, 0, 0, 0)") is "rgba(0, 0, 0, 0)"
FAIL colorTest.parseColor("rgba(204, 0, 102, 0.3)") should be rgba(204, 0, 102, 0.3). Was rgba(204, 0, 102, 0.298039).
PASS colorTest.parseColor("RGBA(255, 255, 255, 0)") is "rgba(255, 255, 255, 0)"
PASS colorTest.parseColor("rgBA(0, 51, 255, 1)") is "rgb(0, 51, 255)"
PASS colorTest.parseColor("rgba(0, 51, 255, 1.1)") is "rgb(0, 51, 255)"
PASS colorTest.parseColor("rgba(0, 51, 255, 37)") is "rgb(0, 51, 255)"
FAIL colorTest.parseColor("rgba(0, 51, 255, 0.42)") should be rgba(0, 51, 255, 0.42). Was rgba(0, 51, 255, 0.419608).
PASS colorTest.parseColor("rgba(0, 51, 255, 0)") is "rgba(0, 51, 255, 0)"
PASS colorTest.parseColor("rgba(0, 51, 255, -0.1)") is "rgba(0, 51, 255, 0)"
PASS colorTest.parseColor("rgba(0, 51, 255, -139)") is "rgba(0, 51, 255, 0)"
FAIL colorTest.parseColor("rgba(42%, 3%, 50%, 0.3)") should be rgba(107, 7, 127, 0.3). Was rgba(107, 7, 128, 0.298039).
PASS colorTest.parseColor("RGBA(100%, 100%, 100%, 0)") is "rgba(255, 255, 255, 0)"
PASS colorTest.parseColor("rgBA(0%, 20%, 100%, 1)") is "rgb(0, 51, 255)"
PASS colorTest.parseColor("rgba(0%, 20%, 100%, 1.1)") is "rgb(0, 51, 255)"
PASS colorTest.parseColor("rgba(0%, 20%, 100%, 37)") is "rgb(0, 51, 255)"
FAIL colorTest.parseColor("rgba(0%, 20%, 100%, 0.42)") should be rgba(0, 51, 255, 0.42). Was rgba(0, 51, 255, 0.419608).
PASS colorTest.parseColor("rgba(0%, 20%, 100%, 0)") is "rgba(0, 51, 255, 0)"
PASS colorTest.parseColor("rgba(0%, 20%, 100%, -0.1)") is "rgba(0, 51, 255, 0)"
PASS colorTest.parseColor("rgba(0%, 20%, 100%, -139)") is "rgba(0, 51, 255, 0)"
PASS colorTest.parseColor("rgba(255, 255, 255, 0%)") is "parse error"
PASS colorTest.parseColor("rgba(10%, 50%, 0, 1)") is "parse error"
PASS colorTest.parseColor("rgba(255, 50%, 0%, 1)") is "parse error"
PASS colorTest.parseColor("rgba(0, 0, 0 0)") is "parse error"
PASS colorTest.parseColor("rgba(0, 0, 0, 0deg)") is "parse error"
PASS colorTest.parseColor("rgba(0, 0, 0, light)") is "parse error"
PASS colorTest.parseColor("rgba()") is "parse error"
PASS colorTest.parseColor("rgba(0)") is "parse error"
PASS colorTest.parseColor("rgba(0, 0, 0)") is "parse error"
PASS colorTest.parseColor("rgba(0, 0, 0, 0, 0)") is "parse error"
PASS colorTest.parseColor("rgba(0%)") is "parse error"
PASS colorTest.parseColor("rgba(0%, 0%)") is "parse error"
PASS colorTest.parseColor("rgba(0%, 0%, 0%)") is "parse error"
PASS colorTest.parseColor("rgba(0%, 0%, 0%, 0%)") is "parse error"
PASS colorTest.parseColor("rgba(0%, 0%, 0%, 0%, 0%)") is "parse error"
PASS colorTest.parseColor("HSL(0, 0%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsL(0, 100%, 50%)") is "rgb(255, 0, 0)"
PASS colorTest.parseColor("hsl(60, 100%, 37.5%)") is "rgb(191, 191, 0)"
PASS colorTest.parseColor("hsl(780, 100%, 37.5%)") is "rgb(191, 191, 0)"
PASS colorTest.parseColor("hsl(-300, 100%, 37.5%)") is "rgb(191, 191, 0)"
PASS colorTest.parseColor("hsl(300, 50%, 50%)") is "rgb(191, 63, 191)"
PASS colorTest.parseColor("hsl(10, 50%, 0)") is "parse error"
PASS colorTest.parseColor("hsl(50%, 50%, 0%)") is "parse error"
PASS colorTest.parseColor("hsl(0, 0% 0%)") is "parse error"
PASS colorTest.parseColor("hsl(30deg, 100%, 100%)") is "parse error"
PASS colorTest.parseColor("hsl(0, 0%, light)") is "parse error"
PASS colorTest.parseColor("hsl()") is "parse error"
PASS colorTest.parseColor("hsl(0)") is "parse error"
PASS colorTest.parseColor("hsl(0, 0%)") is "parse error"
PASS colorTest.parseColor("hsl(0, 0%, 0%, 0%)") is "parse error"
PASS colorTest.parseColor("HSLA(-300, 100%, 37.5%, 1)") is "rgb(191, 191, 0)"
PASS colorTest.parseColor("hsLA(-300, 100%, 37.5%, 12)") is "rgb(191, 191, 0)"
PASS colorTest.parseColor("hsla(-300, 100%, 37.5%, 0.2)") is "rgba(191, 191, 0, 0.2)"
PASS colorTest.parseColor("hsla(-300, 100%, 37.5%, 0)") is "rgba(191, 191, 0, 0)"
PASS colorTest.parseColor("hsla(-300, 100%, 37.5%, -3)") is "rgba(191, 191, 0, 0)"
PASS colorTest.parseColor("hsla(10, 50%, 0, 1)") is "parse error"
PASS colorTest.parseColor("hsla(50%, 50%, 0%, 1)") is "parse error"
PASS colorTest.parseColor("hsla(0, 0% 0%, 1)") is "parse error"
PASS colorTest.parseColor("hsla(30deg, 100%, 100%, 1)") is "parse error"
PASS colorTest.parseColor("hsla(0, 0%, light, 1)") is "parse error"
PASS colorTest.parseColor("hsla()") is "parse error"
PASS colorTest.parseColor("hsla(0)") is "parse error"
PASS colorTest.parseColor("hsla(0, 0%)") is "parse error"
PASS colorTest.parseColor("hsla(0, 0%, 0%, 50%)") is "parse error"
PASS colorTest.parseColor("hsla(0, 0%, 0%, 1, 0%)") is "parse error"
PASS colorTest.parseColor("cmyk(0, 0, 0, 0)") is "parse error"
PASS successfullyParsed is true

TEST COMPLETE
